move each editor to own componenent and make codemirror work
