+++ ux/skill_modules/package.json
@@ -22,6 +22,7 @@
   "devDependencies": {
     "@emotion/react": "^11.10.0",
     "@emotion/styled": "^11.10.0",
+    "@monaco-editor/react": "^4.4.6",
     "@mui/icons-material": "^5.8.4",
     "@mui/material": "^5.10.0",
     "@rollup/plugin-commonjs": "^22.0.2",
@@ -37,6 +38,8 @@
     "@types/leaflet": "^1.9.0",
     "@types/react": "^16.14.30",
     "@types/react-dom": "^16.9.14",
+    "@uiw/react-codemirror": "2.2.1",
+    "codemirror": "5.65.12",
     "react": ">=17.0.2",
     "react-dom": ">=17.0.2",
     "react-scripts": "^5.0.1",
@@ -51,13 +54,11 @@
     "typescript": "^4.1.6"
   },
   "dependencies": {
-    "@monaco-editor/react": "^4.4.6",
     "@types/react-cytoscapejs": "^1.2.2",
     "cytoscape": "^3.22.1",
     "cytoscape-dagre": "^2.4.0",
     "leaflet": "^1.9.3",
     "mui-tree-select": "^1.0.0-beta.9",
-    "react-codemirror2": "^7.2.1",
     "react-cytoscapejs": "^1.2.1",
     "react-leaflet": "^4.2.0",
     "sparnatural": "^8.1.0"
